pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "seccompute"
version = "3.0.0"
description = "Seccomp profile hardening score engine"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
dependencies = [
    "PyYAML>=6.0",
]

[project.optional-dependencies]
viz = []
cli = []
full = ["seccompute[viz,cli]"]
dev = ["pytest>=7.0", "PyYAML>=6.0"]

[project.scripts]
seccompute = "seccompute.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["seccompute*"]

[tool.setuptools.package-data]
seccompute = ["rules/*.yaml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
